services/visualization/requirements.txt
flask==2.3.3
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.17.0
seaborn==0.12.2
redis==5.2.1
python-dotenv==1.0.0
gunicorn==21.2.0 